fix: AttractorVisCanvas component renamed from invalid name + AttractoVisCanvas moved into separate folder
