+++ src/components/react-components/attractor-visualizer/AttractorVisCanvas/AttractorVisCanvas.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef, useState } from 'react';
-import AttractorVisualizer from '../../../services/attractor-visualizer/AttractorVisualizer';
+import AttractorVisualizer from '../../../../services/attractor-visualizer/AttractorVisualizer';
 
-const BifurcationExplorerCanvas: React.FC = () => {
+const AttractorVisCanvas: React.FC = () => {
   const [initialized, setInitialized] = useState<boolean>(false);
 
   const containerRef = useRef<HTMLDivElement | null>(null);
@@ -25,4 +25,4 @@ const BifurcationExplorerCanvas: React.FC = () => {
   );
 };
 
-export default BifurcationExplorerCanvas;
+export default AttractorVisCanvas;
